# Disable all make suffix rules for improved EC2 performance