@colinschmidt 2020-05-28
Is there a way to be sure we weren't relying on implicit rules anywhere?