@davidbiancolin 2020-05-28
> Is there a way to be sure we weren't relying on implicit rules anywhere?

I think that's non-trivial, but i'll look into it. The best course of action imo is either reject this PR, or rely on CI and dog fooding. I don't think this should be included in this release.

I think this is pretty safe, as @a0u has pointed out, they are mostly pretty baroque. The only problematic ones are related to .o, .a i'd imagine.