@jwright6323 2020-05-28
I tried this on BWRC for fun; it goes from 3.2s to 1.9s, although it varies by half a second or so.