@jwright6323 2020-05-28
> I think this is pretty safe, as @a0u has pointed out, they are mostly pretty baroque. The only problematic ones are related to .o, .a i'd imagine.

If it ain't baroque, don't fix it.