@colinschmidt 2020-05-28
Definitely a next release kind of change.
I was just wondering if maybe some verbose flag would print any rules that were reachable by an implicit rule or something like that.