@jwright6323 2020-05-28
`make -Bnd | grep "Found an implicit rule"` seems to give some useful info on a simple C test case:
```
$ make -Bnd | grep "Found an implicit rule"
   Found an implicit rule for 'a.o'.
```

and fails (correctly) when I add `.SUFFIXES`:
```
make -Bnd | grep "Found an implicit rule"
make: *** No rule to make target 'a.o', needed by 'b'.  Stop.
```


In chipyard `sims/vcs`, I get the following with or without `.SUFFIXES:`, so it seems like none of the rules it finds have changed.
```
$ make -Bnd | grep "Found an implicit rule"
  Found an implicit rule for 'IniReader.deppo'.
  Found an implicit rule for 'MultiChannelMemorySystem.deppo'.
  Found an implicit rule for 'CommandQueue.deppo'.
  Found an implicit rule for 'MemoryController.deppo'.
  Found an implicit rule for 'ClockDomain.deppo'.
  Found an implicit rule for 'Bank.deppo'.
  Found an implicit rule for 'SimulatorObject.deppo'.
  Found an implicit rule for 'MemorySystem.deppo'.
  Found an implicit rule for 'Rank.deppo'.
  Found an implicit rule for 'BankState.deppo'.
  Found an implicit rule for 'BusPacket.deppo'.
  Found an implicit rule for 'Transaction.deppo'.
  Found an implicit rule for 'PrintMacros.deppo'.
  Found an implicit rule for 'AddressMapping.deppo'.
  Found an implicit rule for 'IniReader.dep'.
  Found an implicit rule for 'MultiChannelMemorySystem.dep'.
  Found an implicit rule for 'CommandQueue.dep'.
  Found an implicit rule for 'MemoryController.dep'.
  Found an implicit rule for 'ClockDomain.dep'.
  Found an implicit rule for 'Bank.dep'.
  Found an implicit rule for 'TraceBasedSim.dep'.
  Found an implicit rule for 'SimulatorObject.dep'.
  Found an implicit rule for 'MemorySystem.dep'.
  Found an implicit rule for 'Rank.dep'.
  Found an implicit rule for 'BankState.dep'.
  Found an implicit rule for 'BusPacket.dep'.
  Found an implicit rule for 'Transaction.dep'.
  Found an implicit rule for 'PrintMacros.dep'.
  Found an implicit rule for 'AddressMapping.dep'.
  Found an implicit rule for 'IniReader.deppo'.
  Found an implicit rule for 'MultiChannelMemorySystem.deppo'.
  Found an implicit rule for 'CommandQueue.deppo'.
  Found an implicit rule for 'MemoryController.deppo'.
  Found an implicit rule for 'ClockDomain.deppo'.
  Found an implicit rule for 'Bank.deppo'.
  Found an implicit rule for 'SimulatorObject.deppo'.
  Found an implicit rule for 'MemorySystem.deppo'.
  Found an implicit rule for 'Rank.deppo'.
  Found an implicit rule for 'BankState.deppo'.
  Found an implicit rule for 'BusPacket.deppo'.
  Found an implicit rule for 'Transaction.deppo'.
  Found an implicit rule for 'PrintMacros.deppo'.
  Found an implicit rule for 'AddressMapping.deppo'.
  Found an implicit rule for 'IniReader.dep'.
  Found an implicit rule for 'MultiChannelMemorySystem.dep'.
  Found an implicit rule for 'CommandQueue.dep'.
  Found an implicit rule for 'MemoryController.dep'.
  Found an implicit rule for 'ClockDomain.dep'.
  Found an implicit rule for 'Bank.dep'.
  Found an implicit rule for 'TraceBasedSim.dep'.
  Found an implicit rule for 'SimulatorObject.dep'.
  Found an implicit rule for 'MemorySystem.dep'.
  Found an implicit rule for 'Rank.dep'.
  Found an implicit rule for 'BankState.dep'.
  Found an implicit rule for 'BusPacket.dep'.
  Found an implicit rule for 'Transaction.dep'.
  Found an implicit rule for 'PrintMacros.dep'.
  Found an implicit rule for 'AddressMapping.dep'.
   Found an implicit rule for 'AddressMapping.o'.
   Found an implicit rule for 'PrintMacros.o'.
   Found an implicit rule for 'Transaction.o'.
   Found an implicit rule for 'BusPacket.o'.
   Found an implicit rule for 'BankState.o'.
   Found an implicit rule for 'Rank.o'.
   Found an implicit rule for 'MemorySystem.o'.
   Found an implicit rule for 'SimulatorObject.o'.
   Found an implicit rule for 'Bank.o'.
   Found an implicit rule for 'ClockDomain.o'.
   Found an implicit rule for 'MemoryController.o'.
   Found an implicit rule for 'CommandQueue.o'.
   Found an implicit rule for 'MultiChannelMemorySystem.o'.
   Found an implicit rule for 'IniReader.o'.
```