@davidbiancolin 2020-05-28
The only implicit rules of note there are for compiling DRAMSim2 objects, and that happens in a completely separate makefile where the SUFFIX rules still apply.

I was mostly concerned with having to enumerate all the recipes in all of our Makefiles. If we could do that programmatically then we could prove to ourselves this was safe.