@davidbiancolin 2020-05-28
Or, those rules just probably aren't suffix rules.